requirements.txt
python-kraken-sdk
psutil>=5.9.0 